requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
openpyxl>=3.1.0